curelia-platform/backend/requirements.txt
# Curelia Health Platform Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic[email]==2.6.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
sqlalchemy-utils==0.41.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# CORS
starlette==0.27.0

# File handling
python-magic==0.4.27
aiofiles==23.2.1
reportlab==4.1.0
pypdf==4.0.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.1
pydantic-settings==2.0.3
uuid==1.30
pytz==2023.3
email-validator==2.1.0.post1
phonenumbers==8.13.30
geopy==2.4.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.5
pytest-cov==4.1.0
requests==2.31.0

# Documentation
markdown==3.5.1
pyyaml==6.0.1

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Monitoring and logging
structlog==23.2.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk==1.40.4

# Production
gunicorn==21.2.0 